refactor: shared에 ui 폴더로 UI컴포넌트들 이동시킴
- react19 부터는 forwardRef 없이도 props로 ref를 넘길 수 있는 대신 type은 ComponentWithoutRef 와 ComponentProps 구분해서 사용해야함(타입 안전성 확보)
