--- src/entities/index.tsx
@@ -1,272 +1,0 @@
-import * as React from "react"
-import * as SelectPrimitive from "@radix-ui/react-select"
-import * as DialogPrimitive from "@radix-ui/react-dialog"
-import { Check, ChevronDown, X } from "lucide-react"
-import { cva, VariantProps } from "class-variance-authority"
-
-const buttonVariants = cva(
-  "inline-flex items-center justify-center rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:opacity-50 disabled:pointer-events-none ring-offset-background",
-  {
-    variants: {
-      variant: {
-        default: "bg-blue-500 text-white hover:bg-blue-600",
-        destructive: "bg-red-500 text-white hover:bg-red-600",
-        outline: "border border-gray-300 bg-transparent text-gray-700 hover:bg-gray-100",
-        secondary: "bg-gray-200 text-gray-800 hover:bg-gray-300",
-        ghost: "bg-transparent text-gray-700 hover:bg-gray-100",
-        link: "underline-offset-4 hover:underline text-blue-500",
-      },
-      size: {
-        default: "h-10 py-2 px-4",
-        sm: "h-8 px-3 rounded-md text-xs",
-        lg: "h-11 px-8 rounded-md",
-        icon: "h-9 w-9",
-      },
-    },
-    defaultVariants: {
-      variant: "default",
-      size: "default",
-    },
-  },
-)
-
-interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement>, VariantProps<typeof buttonVariants> {
-  className?: string
-}
-
-export const Button = ({ className, variant, size, ...props }: ButtonProps) => {
-  return <button className={buttonVariants({ variant, size, className })} {...props} />
-}
-
-Button.displayName = "Button"
-
-// 입력 컴포넌트
-export interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
-  className?: string
-}
-
-export const Input = ({ className = "", type, ...props }: InputProps) => {
-  return (
-    <input
-      type={type}
-      className={`flex h-10 w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
-      {...props}
-    />
-  )
-}
-Input.displayName = "Input"
-
-// 카드 컴포넌트
-export interface CardProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string
-}
-
-export const Card = ({ className = "", ...props }: CardProps) => (
-  <div className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props} />
-)
-Card.displayName = "Card"
-
-export interface CardHeaderProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string
-}
-
-export const CardHeader = ({ className = "", ...props }: CardHeaderProps) => (
-  <div className={`flex flex-col space-y-1.5 p-6 ${className}`} {...props} />
-)
-CardHeader.displayName = "CardHeader"
-
-export interface CardTitleProps extends React.HTMLAttributes<HTMLHeadingElement> {
-  className?: string
-}
-
-export const CardTitle = ({ className = "", ...props }: CardTitleProps) => (
-  <h3 className={`text-2xl font-semibold leading-none tracking-tight ${className}`} {...props} />
-)
-CardTitle.displayName = "CardTitle"
-
-export interface CardContentProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string
-}
-
-export const CardContent = ({ className = "", ...props }: CardContentProps) => (
-  <div className={`p-6 pt-0 ${className}`} {...props} />
-)
-CardContent.displayName = "CardContent"
-
-// 텍스트 영역 컴포넌트
-export interface TextareaProps extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {
-  className?: string
-}
-
-export const Textarea = ({ className = "", ...props }: TextareaProps) => {
-  return (
-    <textarea
-      className={`flex min-h-[150px] w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
-      {...props}
-    />
-  )
-}
-Textarea.displayName = "Textarea"
-
-// 선택 컴포넌트
-export const Select = SelectPrimitive.Root
-export const SelectGroup = SelectPrimitive.Group
-export const SelectValue = SelectPrimitive.Value
-
-export type SelectTriggerProps = React.ComponentPropsWithoutRef<typeof SelectPrimitive.Trigger> & {
-  className?: string
-}
-
-export const SelectTrigger = ({ className = "", children, ...props }: SelectTriggerProps) => (
-  <SelectPrimitive.Trigger
-    className={`flex h-10 items-center justify-between rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
-    {...props}
-  >
-    {children}
-    <ChevronDown className="h-4 w-4 opacity-50" />
-  </SelectPrimitive.Trigger>
-)
-SelectTrigger.displayName = SelectPrimitive.Trigger.displayName
-
-export type SelectContentProps = React.ComponentPropsWithoutRef<typeof SelectPrimitive.Content> & {
-  className?: string
-  position?: "popper" | "item-aligned"
-}
-
-export const SelectContent = ({ className = "", children, position = "popper", ...props }: SelectContentProps) => (
-  <SelectPrimitive.Portal>
-    <SelectPrimitive.Content
-      className={`relative z-50 min-w-[8rem] overflow-hidden rounded-md border bg-white text-popover-foreground shadow-md data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 ${className}`}
-      position={position}
-      {...props}
-    >
-      <SelectPrimitive.Viewport className="p-1">{children}</SelectPrimitive.Viewport>
-    </SelectPrimitive.Content>
-  </SelectPrimitive.Portal>
-)
-SelectContent.displayName = SelectPrimitive.Content.displayName
-
-export type SelectItemProps = React.ComponentPropsWithoutRef<typeof SelectPrimitive.Item> & {
-  className?: string
-}
-
-export const SelectItem = ({ className = "", children, ...props }: SelectItemProps) => (
-  <SelectPrimitive.Item
-    className={`relative flex w-full cursor-default select-none items-center rounded-sm py-1.5 pl-8 pr-2 text-sm outline-none focus:bg-accent focus:text-accent-foreground data-[disabled]:pointer-events-none data-[disabled]:opacity-50 ${className}`}
-    {...props}
-  >
-    <span className="absolute left-2 flex h-3.5 w-3.5 items-center justify-center">
-      <SelectPrimitive.ItemIndicator>
-        <Check className="h-4 w-4" />
-      </SelectPrimitive.ItemIndicator>
-    </span>
-    <SelectPrimitive.ItemText>{children}</SelectPrimitive.ItemText>
-  </SelectPrimitive.Item>
-)
-SelectItem.displayName = SelectPrimitive.Item.displayName
-
-// 대화상자 컴포넌트
-export const Dialog = DialogPrimitive.Root
-export const DialogTrigger = DialogPrimitive.Trigger
-export const DialogPortal = DialogPrimitive.Portal
-export const DialogOverlay = DialogPrimitive.Overlay
-
-export type DialogContentProps = React.ComponentPropsWithoutRef<typeof DialogPrimitive.Content> & {
-  className?: string
-}
-
-export const DialogContent = ({ className = "", children, ...props }: DialogContentProps) => (
-  <DialogPortal>
-    <DialogOverlay className="fixed inset-0 z-50 bg-black/50 data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0" />
-    <DialogPrimitive.Content
-      className={`fixed left-[50%] top-[50%] z-50 grid w-full max-w-lg translate-x-[-50%] translate-y-[-50%] gap-4 border bg-white p-6 shadow-lg duration-200 data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[state=closed]:slide-out-to-left-1/2 data-[state=closed]:slide-out-to-top-[48%] data-[state=open]:slide-in-from-left-1/2 data-[state=open]:slide-in-from-top-[48%] sm:rounded-lg md:w-full ${className}`}
-      {...props}
-    >
-      {children}
-      <DialogPrimitive.Close className="absolute right-4 top-4 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none data-[state=open]:bg-accent data-[state=open]:text-muted-foreground">
-        <X className="h-4 w-4" />
-        <span className="sr-only">닫기</span>
-      </DialogPrimitive.Close>
-    </DialogPrimitive.Content>
-  </DialogPortal>
-)
-DialogContent.displayName = DialogPrimitive.Content.displayName
-
-export type DialogHeaderProps = React.HTMLAttributes<HTMLDivElement>
-
-export const DialogHeader = ({ className = "", ...props }: DialogHeaderProps) => (
-  <div className={`flex flex-col space-y-1.5 text-center sm:text-left ${className}`} {...props} />
-)
-DialogHeader.displayName = "DialogHeader"
-
-export type DialogTitleProps = React.ComponentPropsWithoutRef<typeof DialogPrimitive.Title> & {
-  className?: string
-}
-
-export const DialogTitle = ({ className = "", ...props }: DialogTitleProps) => (
-  <DialogPrimitive.Title className={`text-lg font-semibold leading-none tracking-tight ${className}`} {...props} />
-)
-DialogTitle.displayName = DialogPrimitive.Title.displayName
-
-// 테이블 컴포넌트
-export interface TableProps extends React.TableHTMLAttributes<HTMLTableElement> {
-  ref?: React.Ref<HTMLTableElement>
-}
-
-export const Table = ({ className = "", ...props }: TableProps) => (
-  <div className="w-full overflow-auto">
-    <table className={`table-fixed w-full caption-bottom text-sm ${className}`} {...props} />
-  </div>
-)
-Table.displayName = "Table"
-
-export interface TableHeaderProps extends React.HTMLAttributes<HTMLTableSectionElement> {
-  className?: string
-}
-
-export const TableHeader = ({ className = "", ...props }: TableHeaderProps) => (
-  <thead className={`[&_tr]:border-b ${className}`} {...props} />
-)
-TableHeader.displayName = "TableHeader"
-
-export interface TableBodyProps extends React.HTMLAttributes<HTMLTableSectionElement> {
-  className?: string
-}
-
-export const TableBody = ({ className = "", ...props }: TableBodyProps) => (
-  <tbody className={`[&_tr:last-child]:border-0 ${className}`} {...props} />
-)
-TableBody.displayName = "TableBody"
-
-export interface TableRowProps extends React.HTMLAttributes<HTMLTableRowElement> {
-  className?: string
-}
-
-export const TableRow = ({ className = "", ...props }: TableRowProps) => (
-  <tr
-    className={`border-b transition-colors hover:bg-muted/50 data-[state=selected]:bg-muted h-14 ${className}`}
-    {...props}
-  />
-)
-TableRow.displayName = "TableRow"
-
-export interface TableHeadProps extends React.ThHTMLAttributes<HTMLTableCellElement> {
-  className?: string
-}
-
-export const TableHead = ({ className = "", ...props }: TableHeadProps) => (
-  <th
-    className={`h-10 px-2 text-left align-middle font-medium text-muted-foreground [&:has([role=checkbox])]:pr-0 ${className}`}
-    {...props}
-  />
-)
-TableHead.displayName = "TableHead"
-
-export interface TableCellProps extends React.TdHTMLAttributes<HTMLTableCellElement> {
-  className?: string
-}
-
-export const TableCell = ({ className = "", ...props }: TableCellProps) => (
-  <td className={`p-2 align-middle [&:has([role=checkbox])]:pr-0 ${className}`} {...props} />
-)
-TableCell.displayName = "TableCell"
