+++ src/shared/ui/Select.tsx
@@ -1,0 +1,60 @@
+import * as SelectPrimitive from "@radix-ui/react-select"
+import { Check, ChevronDown } from "lucide-react"
+
+// 선택 컴포넌트
+export const Select = SelectPrimitive.Root
+export const SelectGroup = SelectPrimitive.Group
+export const SelectValue = SelectPrimitive.Value
+
+// 선택 트리거 컴포넌트: 토글역할을 함
+export type SelectTriggerProps = React.ComponentProps<typeof SelectPrimitive.Trigger>
+
+export const SelectTrigger = ({ className = "", children, ...props }: SelectTriggerProps) => (
+  <SelectPrimitive.Trigger
+    className={`flex h-10 items-center justify-between rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
+    {...props}
+  >
+    {children}
+    <ChevronDown className="h-4 w-4 opacity-50" />
+  </SelectPrimitive.Trigger>
+)
+SelectTrigger.displayName = SelectPrimitive.Trigger.displayName
+
+// 선택 컨텐츠 컴포넌트
+export type SelectContentProps = React.ComponentProps<typeof SelectPrimitive.Content> & {
+  className?: string
+  position?: "popper" | "item-aligned"
+}
+
+export const SelectContent = ({ className = "", children, position = "popper", ...props }: SelectContentProps) => (
+  <SelectPrimitive.Portal>
+    <SelectPrimitive.Content
+      className={`relative z-50 min-w-[8rem] overflow-hidden rounded-md border bg-white text-popover-foreground shadow-md data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 ${className}`}
+      position={position}
+      {...props}
+    >
+      <SelectPrimitive.Viewport className="p-1">{children}</SelectPrimitive.Viewport>
+    </SelectPrimitive.Content>
+  </SelectPrimitive.Portal>
+)
+SelectContent.displayName = SelectPrimitive.Content.displayName
+
+// 선택 아이템 컴포넌트
+export type SelectItemProps = React.ComponentProps<typeof SelectPrimitive.Item> & {
+  className?: string
+}
+
+export const SelectItem = ({ className = "", children, ...props }: SelectItemProps) => (
+  <SelectPrimitive.Item
+    className={`relative flex w-full cursor-default select-none items-center rounded-sm py-1.5 pl-8 pr-2 text-sm outline-none focus:bg-accent focus:text-accent-foreground data-[disabled]:pointer-events-none data-[disabled]:opacity-50 ${className}`}
+    {...props}
+  >
+    <span className="absolute left-2 flex h-3.5 w-3.5 items-center justify-center">
+      <SelectPrimitive.ItemIndicator>
+        <Check className="h-4 w-4" />
+      </SelectPrimitive.ItemIndicator>
+    </span>
+    <SelectPrimitive.ItemText>{children}</SelectPrimitive.ItemText>
+  </SelectPrimitive.Item>
+)
+SelectItem.displayName = SelectPrimitive.Item.displayName
